moving test into /messaging
