--- com.unity.netcode.gameobjects/Tests/Runtime/DisconnectReasonTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93141fa15824f406b89dbc6f32c8910d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
