+++ com.unity.netcode.gameobjects/Tests/Runtime/Messaging/DisconnectReasonTests.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections;
+using System.Text.RegularExpressions;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -24,17 +26,23 @@ namespace Unity.Netcode.RuntimeTests
         }
 
         private int m_DisconnectCount;
+        private bool m_ThrowOnDisconnect = false;
 
         public void OnClientDisconnectCallback(ulong clientId)
         {
             m_DisconnectCount++;
-            Debug.Log($"Disconnected {clientId}");
+            if (m_ThrowOnDisconnect)
+            {
+                throw new SystemException("whatever");
+            }
         }
 
         [UnityTest]
         public IEnumerator DisconnectReasonTest()
         {
             float startTime = Time.realtimeSinceStartup;
+            m_ThrowOnDisconnect = false;
+            m_DisconnectCount = 0;
 
             // Add a callback for first client, when they get disconnected
             m_ClientNetworkManagers[0].OnClientDisconnectCallback += OnClientDisconnectCallback;
@@ -51,6 +59,35 @@ namespace Unity.Netcode.RuntimeTests
 
             Assert.AreEqual(m_ClientNetworkManagers[0].DisconnectReason, "Bogus reason 1");
             Assert.AreEqual(m_ClientNetworkManagers[1].DisconnectReason, "Bogus reason 2");
+
+            Debug.Assert(m_DisconnectCount == 2);
+        }
+
+        [UnityTest]
+        public IEnumerator DisconnectExceptionTest()
+        {
+            m_ThrowOnDisconnect = true;
+            m_DisconnectCount = 0;
+            float startTime = Time.realtimeSinceStartup;
+
+            // Add a callback for first client, when they get disconnected
+            m_ClientNetworkManagers[0].OnClientDisconnectCallback += OnClientDisconnectCallback;
+            m_ClientNetworkManagers[1].OnClientDisconnectCallback += OnClientDisconnectCallback;
+
+            // Disconnect first client, from the server
+            LogAssert.Expect(LogType.Exception, new Regex(".*whatever.*"));
+            m_ServerNetworkManager.DisconnectClient(m_ClientNetworkManagers[0].LocalClientId);
+
+            // Disconnect second client, from the server
+            LogAssert.Expect(LogType.Exception, new Regex(".*whatever.*"));
+            m_ServerNetworkManager.DisconnectClient(m_ClientNetworkManagers[1].LocalClientId);
+
+            while (m_DisconnectCount < 2 && Time.realtimeSinceStartup < startTime + 10.0f)
+            {
+                yield return null;
+            }
+
+            Debug.Assert(m_DisconnectCount == 2);
         }
     }
 }
